requirements.txt
datetime
requests
Pillow
openai
xmltodict